+++ src/fit/include/RAT/BONSAI/timefit.h
@@ -8,7 +8,7 @@
 // **********************************************
 
 class timefit {
-  short int bgfit;
+  int bgfit;
   unsigned int nlike, *offset, *nneg, *lmax;
   unsigned int *loglike, *time_histo;
   float *qminmax, *pdf, *pdf_sum, *pdf_max;
@@ -48,11 +48,11 @@ class timefit {
   timefit(hitsel *hs);
   ~timefit();
 
-  inline float fittime(short int fast, float *vertex, float &dt0);
+  inline float fittime(int fast, float *vertex, float &dt0);
   void fittime(int ndt, float *dts, float &t0, float &dt);
-  float fittime(short int fast, float *vertex, float *direct, float &dt);
+  float fittime(int fast, float *vertex, float *direct, float &dt);
   float fittime(float *vertex, float *direct, float &dt, float tmi, float tpl);
-  float fittime_lcor(short int fast, float *vertex, float *direct, float &dt);
+  float fittime_lcor(int fast, float *vertex, float *direct, float &dt);
   float chargetest(float *vertex, float *direct, float &qin, float &qoutplus, float *qoutminus);
   float maxconecharge(float *vert, float *direct);
   float fittime(float *vertex, float &dt0, float &signal);
@@ -64,7 +64,7 @@ class timefit {
   float ntgood(float *vertex, float bgrate, float &guncor);
   float shapetest(void);
   inline int sets(void);
-  inline int gethits(short int fast, float *vertex, int *cabs, float *tofs, float *dirs, float *weight);
+  inline int gethits(int fast, float *vertex, int *cabs, float *tofs, float *dirs, float *weight);
   inline void getdist(int set, float t0, int nbin, float tmin, float tmax, float *cont);
   inline float *chargebins(void);
 };
